Web app/requirements.txt
flask>=2.3
flask-cors>=4.0
tensorflow>=2.12
numpy>=1.24
Pillow>=10.0
opencv-python>=4.8
